refactor intake route modules
