+++ src/app/router_registry.py
@@ -98,6 +98,8 @@ from app.routers.dpm_wave_evidence import router as dpm_wave_evidence_router
 from app.routers.dpm_waves import router as dpm_waves_router
 from app.routers.foundation import router as foundation_router
 from app.routers.intake import router as intake_router
+from app.routers.intake_uploads import router as intake_uploads_router
+from app.routers.lookup_catalogs import router as lookup_catalogs_router
 from app.routers.platform import router as platform_router
 from app.routers.portfolio import router as portfolio_router
 from app.routers.portfolio_activity import router as portfolio_activity_router
@@ -172,6 +174,8 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(domain_products_router)
     app.include_router(source_products_router)
     app.include_router(intake_router)
+    app.include_router(intake_uploads_router)
+    app.include_router(lookup_catalogs_router)
     app.include_router(foundation_router)
     app.include_router(portfolio_router)
     app.include_router(portfolio_activity_router)
